+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.2] - 2024-04-08
+
+### Fixed
+- `IdenticalTransformOverrides` now also checks for individual euler angles. (-0, 90, -0) == (0, 90, 0)
+
 ## [1.3.1] - 2024-03-25
 
 ### Fixed